requirements.txt
numpy
opencv-python
pillow
pixeloe